# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyBusiness 2.0 API",
        "description": "MyBusiness API documentation",
        "contact": {
            "email": "sgg.2020@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/course/index": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get all course",
                "operationId": "indexCourse",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/course/status/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get courses by status",
                "operationId": "showCoursesByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Boolean ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/course/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Get course by ID",
                "operationId": "showCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "String ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/course/add": {
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Create course",
                "operationId": "cf53a814b952d1e13ce14c6ec02239ce",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "code": {
                                                "type": "string"
                                            },
                                            "references_number": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "department_id": {
                                                "type": "uuid"
                                            },
                                            "program_id": {
                                                "type": "uuid"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "code": "MSG",
                                    "references_number": "87654",
                                    "name": "Master",
                                    "description": "Degree details",
                                    "department_id": "e5040c40-0a22-469b-8194-34f722aeffc2",
                                    "program_id": "e5040c40-0a22-469b-8194-34f722aeffc2",
                                    "active": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/course/update/{id}": {
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "Update course",
                "operationId": "71542fa3486925626534bc1881664162",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "uuid",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "code": {
                                                "type": "string"
                                            },
                                            "references_number": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "department_id": {
                                                "type": "uuid"
                                            },
                                            "program_id": {
                                                "type": "uuid"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "code": "MSG",
                                    "references_number": "87654",
                                    "name": "Master",
                                    "description": "Degree details",
                                    "department_id": "e5040c40-0a22-469b-8194-34f722aeffc2",
                                    "program_id": "e5040c40-0a22-469b-8194-34f722aeffc2",
                                    "active": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/course/delete/{id}": {
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "Delete course by id",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "0C6A36BA-10E4-438F-BA86-0D5B68A2BB15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}